Updated main

+++ services/PrintAnalyzer.cpp
@@ -34,12 +34,13 @@ vector<Disease *> PrintAnalyzer::Analyze(Print & p)
 {
     DiseaseManager * dism = DiseaseManager::Get();
     vector<Disease *> res;
-    vector<Disease * > vd = dism->GetDiseases();
-    for (int i=0; i< vd.size(); i++) {
-        int compt = 0;
+    vector<Disease *> vd = dism->GetDiseases();
+    for (unsigned int i = 0; i < vd.size(); i++) {
+        unsigned int compt = 0;
         vector<Measurement *> vme = vd[i]->GetMeasurements();
         compt += vme[i]->Comparison(p);
-        if ( compt >= 0.75*vd.size() ){
+        if (compt >= (unsigned int) 0.75 * vd.size())
+		{
             res.push_back(vd[i]);
         }
     }
